refactor: 쿠폰 패키지 이동
