+++ src/main/java/coupon/domain/coupon/CouponName.java
@@ -1,4 +1,4 @@
-package coupon.domain;
+package coupon.domain.coupon;
 
 import lombok.Getter;
