+++ src/main/java/coupon/domain/coupon/DiscountPolicy.java
@@ -1,4 +1,4 @@
-package coupon.domain;
+package coupon.domain.coupon;
 
 public interface DiscountPolicy {
