+++ src/main/java/coupon/domain/coupon/DiscountPolicyViolationException.java
@@ -1,4 +1,4 @@
-package coupon.domain;
+package coupon.domain.coupon;
 
 public class DiscountPolicyViolationException extends RuntimeException {
